Compact MRU recents: pack emoji into 16 bits
Shrinks the MRU storage and split-sync payload by storing each emoji recent
as a 16-bit code (4-bit category | 12-bit glyph offset into
EMJ_CATEGORIES[cat].codepoints[]) instead of a raw 32-bit codepoint.
Languages stay 1-byte LANG_* indices. The emoji layer owns pack/decode; the
"Preset" key resolves its codepoint list to codes by scanning the table.

This is safe because the EEPROM is reset on every firmware update
(build-date magic + EECONFIG size change), so a stored code is always
interpreted against the same emoji table that produced it.

Result:
- split-sync payload (mru_sync_t): 64 -> 40 bytes
- EEPROM user datablock: EECONFIG_USER_DATA_SIZE 92 -> 66
  (MRU adds 36 B: uint16 mru_emoji[12]=24 + uint8 mru_lang[12]=12, no padding)
- empty sentinels: 0xFFFF (emoji code), 0xFF (lang index)

Verified: split72:default and corne42:default both compile to .uf2.

+++ keyboards/handwired/polykybd/split_sync.h
@@ -52,7 +52,7 @@ typedef struct _overlay_map_sync_t {
 // top row identically. Sent only when the lists change (mru_sync_pending()).
 typedef struct _mru_sync_t {
     uint32_t crc32;
-    uint32_t emoji[MRU_CAP];
+    uint16_t emoji[MRU_CAP];
     uint8_t  lang[MRU_CAP];
 } mru_sync_t;
